agent_backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
twilio
sqlmodel
aiosqlite
